boxing-analyst/requirements.txt
streamlit
yt-dlp
google-generativeai
python-dotenv
